requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
distro==1.9.0
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
openai==1.65.2
pillow==10.4.0
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.32.3
resend==2.6.0
sniffio==1.3.1
sqlparse==0.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.3.0
